Convert api to return native types
suds.sudsobject is seriously magic.  More specifically,
it is very difficult to pickle/serialize these objects.
It seems like I'm not the only one running into this problem.
To aid in testing (so I don't need to talk to a real lab lamager
server), being able to serialize the return values from the
api definitely helps.

Now all the return values are converted to either a dict or a
list of dicts.  This required me to update labmanager.shell
to accomodate these new types, but now the api is much easier
to use with less magic going on.

Also, it seems as though I screwed up the library vs. workspace
enums, and fixed those while I was in labmanager.api.