Add checkout/delete commands.
This allows you to go full circle with checking out
a config from the library, deploying it, seeing IPs
for the machines, undeploying it, then deleting it
from the workspace.  May have to use the internal
API for the checkout configuration api call because
it won't work if there are workspaces with the same
name (even if it's across different organizations).