Add documentation
Updated the readme file for an overview of how to
use lmsh.  Documented the cmd line options when
creating the parser.  Started adding documentation
to each command (accessible via "help <command>"),
but I stopped as it doesn't dedent the output.  I will
probably have to write a custom do_help method to
account for this.