Add proper error handling
When an invalid SOAP call occurs, don't
exist the shell, just print the error message
and go back to the prompt.  However, for
a single command, e.g "lmsh list" I want
a nonzero error code when an error occurs.

When suds produces an error, it tries to
log an error message but complains about not
having any handlers.  Given that the error message
logs the raw request (which contains the user passwords),
I did not want that to get logged to the console so
I added a NullHandler to the suds logger.